+++ Assets/Resources/Font/NotoSansKR-Medium SDF_Bold.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 759d703d41ba942228cb3f5bfd186225
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
