+++ Assets/Resources/Prefabs/Panel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b1295d619e1294af9a8da72e730fd99f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
