+++ Assets/Resources/Shader/AlphaToColor.shader
@@ -1,0 +1,130 @@
+Shader "UI/AlphaToColor"
+{
+    Properties
+    {
+        [PerRendererData] _MainTex ("Sprite Texture", 2D) = "white" {}
+        _Color ("Tint", Color) = (1,1,1,1)
+        _WobblesPerSecond("Wobbles per Second ", Float) = 2
+        _WobblePeriod("Wobble Period (pixels)", Float) = 45
+        _WobbleAmplitude("Wobble Amplitude (pixels)", Float) = 3.2
+        _StencilComp ("Stencil Comparison", Float) = 8
+        _Stencil ("Stencil ID", Float) = 0
+        _StencilOp ("Stencil Operation", Float) = 0
+        _StencilWriteMask ("Stencil Write Mask", Float) = 255
+        _StencilReadMask ("Stencil Read Mask", Float) = 255
+
+        _ColorMask ("Color Mask", Float) = 15
+
+        [Toggle(UNITY_UI_ALPHACLIP)] _UseUIAlphaClip ("Use Alpha Clip", Float) = 0
+    }
+
+    SubShader
+    {
+        Tags
+        {
+            "Queue"="Transparent"
+            "IgnoreProjector"="True"
+            "RenderType"="Transparent"
+            "PreviewType"="Plane"
+            "CanUseSpriteAtlas"="True"
+        }
+
+        Stencil
+        {
+            Ref [_Stencil]
+            Comp [_StencilComp]
+            Pass [_StencilOp]
+            ReadMask [_StencilReadMask]
+            WriteMask [_StencilWriteMask]
+        }
+
+        Cull Off
+        Lighting Off
+        ZWrite Off
+        ZTest [unity_GUIZTestMode]
+        Blend SrcAlpha OneMinusSrcAlpha
+        ColorMask [_ColorMask]
+
+        Pass
+        {
+            Name "Default"
+            CGPROGRAM
+            #pragma vertex vert
+            #pragma fragment frag
+            #pragma target 2.0
+
+            #include "UnityCG.cginc"
+            #include "UnityUI.cginc"
+
+            #pragma multi_compile_local _ UNITY_UI_CLIP_RECT
+            #pragma multi_compile_local _ UNITY_UI_ALPHACLIP
+
+            struct appdata_t
+            {
+                float4 vertex : POSITION;
+                float4 color : COLOR;
+                float2 texcoord : TEXCOORD0;
+                UNITY_VERTEX_INPUT_INSTANCE_ID
+            };
+
+            struct v2f
+            {
+                float4 vertex : SV_POSITION;
+                fixed4 color : COLOR;
+                float2 texcoord : TEXCOORD0;
+                float4 worldPosition : TEXCOORD1;
+                UNITY_VERTEX_OUTPUT_STEREO
+            };
+
+            sampler2D _MainTex;
+            fixed4 _Color;
+            fixed4 _TextureSampleAdd;
+
+            /*1*/float _WobblesPerSecond;
+            /*2*/float _WobblePeriod;
+            /*3*/float _WobbleAmplitude;
+            /*4*/#define TWO_PI 6.28318
+            
+            float4 _ClipRect;
+            float4 _MainTex_ST;
+
+            v2f vert(appdata_t IN)
+            {
+                v2f OUT;
+                /*1*/
+                float4 worldPos = IN.vertex;
+                /*2*/
+                float offsetInput = (_Time.y + frac(worldPos.x)) * _WobblesPerSecond * TWO_PI + worldPos.x /
+                    _WobblePeriod;
+                /*3*/
+                worldPos.y += sin(offsetInput) * _WobbleAmplitude;
+                /*4*/
+                worldPos.x += cos(offsetInput) * _WobbleAmplitude;
+                /*5*/
+                OUT.vertex = UnityObjectToClipPos(worldPos);
+                OUT.texcoord = IN.texcoord;
+                #ifdef UNITY_HALF_TEXEL_OFFSET
+                OUT.vertex.xy += (_ScreenParams.zw - 1.0)*float2(-1,1);
+                #endif
+                OUT.color = IN.color * _Color;
+                return OUT;
+            }
+
+            fixed4 frag(v2f IN) : SV_Target
+            {
+                half4 color = (tex2D(_MainTex, IN.texcoord) + _TextureSampleAdd).a * IN.color;
+
+                #ifdef UNITY_UI_CLIP_RECT
+                color.a *= UnityGet2DClipping(IN.worldPosition.xy, _ClipRect);
+                #endif
+
+                #ifdef UNITY_UI_ALPHACLIP
+                clip (color.a - 0.001);
+                #endif
+
+                return color;
+            }
+            ENDCG
+        }
+    }
+}
