+++ Assets/Resources/Shader/AlphaToColor.shader.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: ea9b4a515643445a8a6ec0b379d088e4
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  preprocessorOverride: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
