--- Assets/Resources/Shader/AlphaToColor.shadergraph.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: bc308e5b3865c4bfe8fa931d2d2f3bd1
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 625f186215c104763be7675aa2d941aa, type: 3}
