--- Assets/Resources/Shader/New Material.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a92b921ebef0c42729f93bde3cb55506
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
