--- Assets/Scripts/CircleLayoutGroup.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1120eec4f6d72402a902e65cefba60fa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
