+++ Assets/Scripts/UI/TMPWobble.cs
@@ -1,0 +1,55 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using TMPro;
+
+[ExecuteInEditMode]
+public class TMPWobble : MonoBehaviour
+{
+    private const float Speed  = 1;
+    private const float Strength = 0.5f;
+    
+    private TMP_Text textMesh;
+
+    private Mesh mesh;
+
+    private Vector3[] vertices;
+
+    private float time = 0;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        textMesh = GetComponent<TMP_Text>();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        time += Time.deltaTime * Speed;
+        
+        textMesh.ForceMeshUpdate();
+        mesh = textMesh.mesh;
+        vertices = mesh.vertices;
+
+        for (int i = 0; i < textMesh.textInfo.characterCount; i++)
+        {
+            TMP_CharacterInfo c = textMesh.textInfo.characterInfo[i];
+
+            int index = c.vertexIndex;
+
+            Vector3 offset = Wobble(time + i) * Strength;
+            vertices[index] += offset;
+            vertices[index + 1] += offset;
+            vertices[index + 2] += offset;
+            vertices[index + 3] += offset;
+        }
+
+        mesh.vertices = vertices;
+        textMesh.canvasRenderer.SetMesh(mesh);
+    }
+
+    Vector2 Wobble(float time) {
+        return new Vector2(Mathf.Sin(time*3.3f), Mathf.Cos(time*2.5f));
+    }
+}
