+++ Assets/Scripts/UI/TMPWobble.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4b0678deab1bf403c8c8ad22bfde6ed9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
