--- Assets/TextMesh Pro/Glyph Report.txt
@@ -1,12 +1,0 @@
-Font: Noto Sans KR  Style: Medium
-Point Size: 16   SP/PD Ratio: 31.3%
-
-Characters included: 11932/11932
-Missing characters: 0
-Excluded characters: 0
-
-Characters missing from font file:
-----------------------------------------
-
-Characters excluded from packing:
-----------------------------------------
