--- Assets/TextMesh Pro/Glyph Report.txt.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ab63dac52d5a347a1b12c88008c28248
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
